Refact : <기성 #45> 인증 로직 분리
